Database set up, route connection working:
