+++ Markobacon_Backend/.env-template
@@ -1,1 +1,4 @@
-## TODO
+SITE_ORIGIN=#origin of site ip
+PORT=#port being used for the backend
+URI_USER=#username to connect to the mongodb cluster
+URI_PASS=#password to connect to the mongodb cluster
