+++ Markobacon_Backend/database.js
@@ -1,4 +1,5 @@
 const mongoose = require('mongoose');
+require('dotenv').config()
 
 // Schema(s)
 const Post = new mongoose.Schema({
@@ -19,10 +20,11 @@ mongoose.model('Post', Post);
 mongoose.model('Login', Login)
 
 // Connect to the database
-const uri = `ongodb+srv://${process.env.URI_USER}:${process.env.URI_PASS}@markside.rq6p8.mongodb.net/myFirstDatabase?retryWrites=true&w=majority`;
-mongoose.connect(uri, { newUrlParser: true, useUnifiedTopology: true });
+const uri = `mongodb+srv://${process.env.URI_USER}:${process.env.URI_PASS}@markside.rq6p8.mongodb.net/myFirstDatabase?retryWrites=true&w=majority`;
+mongoose.connect(uri);
 
 // Log connection (in case of error)
 const connection = mongoose.connection;
+console.log(`user is ${process.env.URI_USER} pass is ${process.env.URI_PASS}`)
 connection.on('error', (err) => console.log('Connection failed!\n'+ err))
 connection.once('open', () => console.log('Connected to Markside database successfully'));
